feat: refactored scorecard component
